Beregne etterbetaling fra responsen

+++ src/main/kotlin/no/nav/familie/oppdrag/simulering/SimulerResultatDto.kt
@@ -1,3 +1,3 @@
 package no.nav.familie.oppdrag.simulering
 
-data class SimulerResultatDto(val etterbetaling: Long)
+data class SimulerResultatDto(val etterbetaling: Int)
